+++ src/demo/Demo.vue
@@ -557,8 +557,6 @@ export default {
   },
   mounted() {
     this.startActions();
-
-    console.log(this.$refs.pie.chart);
   }
 };
 </script>