+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Changed `updateOptions.lazyUpdate` to `true` by default. ([#533](https://github.com/ecomfe/vue-echarts/issues/533#issuecomment-809883909))
 * Only perform an additional `resize` call after init within a task. ([#533](https://github.com/ecomfe/vue-echarts/issues/533#issuecomment-809883909))
+* The `.chart` getter API now works for Vue 2. (#542)
 
 ## 6.0.0-rc.4